Refactor: 홈페이지 css 세부 위치 수정
